# swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Messenger API"
host: "messenger.fafago-dev.alexgd.es"
basePath: "/v1"
...
schemes:
- "http"
paths:
  /messenger/room:
    get:
      tags:
      - "messenger"
      summary: "Get all rooms from user logged"
      operationId: "getRooms"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        200:
          description: A room object
          schema:
            type: object
            properties:
              roomId:
                type: integer
                description: The room id
              roomName:
                type: string
                description: The room name
              lastMessage:
                type: string
                description: Thelast message
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header